+++ Bootstrap.php
@@ -26,7 +26,7 @@ class Shopware_Plugins_Frontend_jtlconnector_Bootstrap extends Shopware_Componen
 
     public function getVersion()
     {
-        return '1.1.0';
+        return '1.1.1';
     }
 
     public function getInfo()
@@ -169,6 +169,8 @@ class Shopware_Plugins_Frontend_jtlconnector_Bootstrap extends Shopware_Componen
             case '1.0.11':
                 break;
             case '1.0.12':
+                break;
+            case '1.1.0':
                 break;
             default:
                 return false;