+++ src/jtl/Connector/Shopware/Controller/Manufacturer.php
@@ -6,6 +6,7 @@
 
 namespace jtl\Connector\Shopware\Controller;
 
+use jtl\Connector\Model\Identity;
 use \jtl\Connector\Result\Action;
 use \jtl\Connector\Core\Rpc\Error;
 use \jtl\Connector\Core\Model\QueryFilter;
@@ -13,6 +14,7 @@ use \jtl\Connector\Core\Utilities\DataConverter;
 use \jtl\Connector\Shopware\Utilities\Mmc;
 use \jtl\Connector\Core\Logger\Logger;
 use \jtl\Connector\Formatter\ExceptionFormatter;
+use \jtl\Connector\Core\Utilities\Language as LanguageUtil;
 
 /**
  * Manufacturer Controller
@@ -42,7 +44,21 @@ class Manufacturer extends DataController
                     $manufacturer = Mmc::getModel('Manufacturer');
                     $manufacturer->map(true, DataConverter::toObject($manufacturerSW));
 
+                    // ManufacturerI18n
                     $this->addPos($manufacturer, 'addI18n', 'ManufacturerI18n', $manufacturerSW);
+                    if (isset($manufacturerSW['translations'])) {
+                        foreach ($manufacturerSW['translations'] as $localeName => $translation) {
+                            $manufacturerI18n = Mmc::getModel('ManufacturerI18n');
+                            $manufacturerI18n->setLanguageISO(LanguageUtil::map($localeName))
+                                ->setManufacturerId(new Identity($manufacturerSW['id']))
+                                ->setDescription($translation['description'])
+                                ->setMetaDescription($translation['metaDescription'])
+                                ->setMetaKeywords($translation['metaKeywords'])
+                                ->setTitleTag($translation['metaTitle']);
+
+                            $manufacturer->addI18n($manufacturerI18n);
+                        }
+                    }
 
                     $result[] = $manufacturer;
                 } catch (\Exception $exc) {